feat(api): add OnlyOneOf decorator for validation and refactor CreateNotificationDto
