+++ apps/api/src/modules/notifications/dtos/create-notification.dto.ts
@@ -1,35 +1,6 @@
 import { IsNumber, IsObject, IsString, ValidateIf } from 'class-validator';
 import { IsDataValid } from 'src/common/decorators/is-data-valid.decorator';
-import { registerDecorator, ValidationOptions, ValidationArguments } from 'class-validator';
-
-// Custom decorator to ensure only one of providerId or providerChain is set
-export function OnlyOneOf(
-  property1: string,
-  property2: string,
-  validationOptions?: ValidationOptions,
-) {
-  return function (object: object, propertyName: string) {
-    registerDecorator({
-      name: 'onlyOneOf',
-      target: object.constructor,
-      propertyName: propertyName,
-      options: validationOptions,
-      constraints: [property1, property2],
-      validator: {
-        validate(_: unknown, args: ValidationArguments) {
-          const [prop1, prop2] = args.constraints;
-          const val1 = (args.object as unknown)[prop1];
-          const val2 = (args.object as unknown)[prop2];
-          return (val1 && !val2) || (!val1 && val2);
-        },
-        defaultMessage(args: ValidationArguments) {
-          const [prop1, prop2] = args.constraints;
-          return `Either "${prop1}" or "${prop2}" must be provided, but not both.`;
-        },
-      },
-    });
-  };
-}
+import { OnlyOneOf } from 'src/common/decorators/only-one-of.decorator';
 
 export class CreateNotificationDto {
   // Used to trigger OnlyOneOf validator
